+++ extension/src/content-script/osu/pages/topics.ts
@@ -4,12 +4,27 @@ import { addFlagUsers as addFlagUsers } from "@src/content-script/osu/flagHtml";
 import { TFlagItems } from "@src/utils/html";
 import { nextAbortControllerSignal } from "@src/utils/fetchUtils";
 
+const topicsLoadedObserver = new MutationObserver((mutations) => {
+    const newPosts = [...mutations[0].addedNodes].filter(a => a.nodeType === 1).map(a => (a as Element).querySelector(".forum-post-info")).filter(post => post !== null) as Element[];
+    return addFlagsPostsTopics(newPosts);
+});
+
 export const updateFlagsTopics = async () => {
     const url = location.href;
     if (!url.includes("osu.ppy.sh/community/forums/topics/")) return;
     const signal = nextAbortControllerSignal();
-    const posts = document.querySelectorAll(".forum-post-info");
 
+    const page = document.querySelector(".osu-page--forum-topic");
+    if(!page) return;
+
+
+    topicsLoadedObserver.observe(page, {childList: true});
+
+    const posts = page.querySelectorAll(".forum-post-info");
+    return addFlagsPostsTopics(posts, signal);
+};
+
+const addFlagsPostsTopics = async (posts: Iterable<Element>, signal?: AbortSignal) => {
     const flagItems: TFlagItems = [];
 
     for (const item of posts) {
@@ -24,4 +39,4 @@ export const updateFlagsTopics = async () => {
         addSuperParentClone: true,
         signal: signal,
     });
-};
+}