+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Version 3.3.0
 
 - Flags in osu wiki page
+- Fix topics flags loading more content
 - Flags in wybin home team
 - Flags in wybin home player
 - Flags in wybin tournament edit / staff